Establishing Gallery
First attempts at creating the template for the gallery page. It is not too far from finished at this point, but it does need some additional things. I'll also have to make a JS filter system with a dropdown menu, so there is that. Hopefully, it won't be too hard, and I can get on with my life rather easily.