Minor Changes
I noticed some minor discrepancies with the pages while viewing them on my phone. The Gallery page was displaying the active link (current page) as being the Home page, so I removed the active class from the home link (gallery.html, line 17 & 32) and added it to the gallery link (gallery.html, line 19 & 34).

The other minor changes are to the background and banner images of the pages for Home, About, and Gallery. I duplicated the backdrop of the splash page to be for the Home page for when people visit the site through the home page and such. Since the images are a series (from Horn Island 34), I wanted them to go in order from page to page (4 images, 4 pages).