Skim and Trim
I did a lot in this one. There's a lot that I might miss out on explaining. The gist of it is that I removed access to two pages, combined the idea of the contact page with the about me page, changed the design of the about me page so it works better on all screens, and started messing around with a darker theme. The navigation bar has yet to be updated to match the theme, but it still functions and color will be more of a minor change to it.