Gallery and Structure Update
Change the file folder structure to have all css files in the css folder and all js files in their own folder.

The image gallery is currently further along with this update, as it has a "light box" modal slideshow thing now. There's too many names for it. It be crazy. Anyway, it's not just a bland lifeless page of boxes. It's now a page of boxes and images that need to be re-styled.