Finalizing for In Class Critique
The final variant of the website for class critique. A lot of changes have been made, and the splash page has been scrapped in favor of the home page being the first to show. I simply had too many problems with the splash page for it to be of any real use right now.